set up reorg
