+++ pkg/core/server/grpc.go
@@ -1,0 +1,2 @@
+// implementation of the grpc service definition found in the audius protocol.proto spec
+package server
