+++ pkg/core/server/registry_bridge.go
@@ -1,0 +1,2 @@
+// Keeps the validators updated in cometbft and core up to date with what is present on the ethereum node registry.
+package server
